--- reprezen-swagger-parser/src/main/java/com/reprezen/swaggerparser/impl3/ServerVariableImpl.java
@@ -1,148 +1,0 @@
-package com.reprezen.swaggerparser.impl3;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.reprezen.swaggerparser.impl3.SwaggerObjectImpl;
-import com.reprezen.swaggerparser.jsonoverlay.JsonOverlay;
-import com.reprezen.swaggerparser.jsonoverlay.JsonOverlayFactory;
-import com.reprezen.swaggerparser.jsonoverlay.coll.ValListOverlay;
-import com.reprezen.swaggerparser.jsonoverlay.coll.ValMapOverlay;
-import com.reprezen.swaggerparser.jsonoverlay.std.AnyObjectOverlay;
-import com.reprezen.swaggerparser.jsonoverlay.std.StringOverlay;
-import com.reprezen.swaggerparser.model3.ServerVariable;
-import java.util.Collection;
-import java.util.Map;
-import javax.annotation.Generated;
-
-public class ServerVariableImpl extends SwaggerObjectImpl implements ServerVariable {
-
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public ServerVariableImpl(String key, JsonNode json, JsonOverlay<?> parent) {
-        super(key, json, parent);
-    }
-
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public ServerVariableImpl(String key, JsonOverlay<?> parent) {
-        super(key, parent);
-    }
-
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    private ValListOverlay<String, StringOverlay> enumValues = registerField("enum", "enumValues", null, new ValListOverlay<String, StringOverlay>("enum", this, StringOverlay.factory));;
-
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    private StringOverlay defaultValue = registerField("default", "defaultValue", null, new StringOverlay("default", this));
-
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    private StringOverlay description = registerField("description", "description", null, new StringOverlay("description", this));
-
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    private ValMapOverlay<Object,AnyObjectOverlay> extensions = registerField("", "extensions", "x-.*", new ValMapOverlay<Object, AnyObjectOverlay>("", this, AnyObjectOverlay.factory, "x-.*"));
-
-    // EnumValue
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public Collection<String> getEnumValues() {
-        return enumValues.get();
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public String getEnumValue(int index) {
-        return enumValues.get(index);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setEnumValues(Collection<String> enumValues) {
-        this.enumValues.set((Collection<String>) enumValues);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setEnumValue(int index, String enumValue) {
-        enumValues.set(index, enumValue);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void addEnumValue(String enumValue) {
-        enumValues.add(enumValue);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void removeEnumValue(int index) {
-        enumValues.remove(index);
-    }
-
-    // Default
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public String getDefault() {
-        return defaultValue.get();
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setDefault(String defaultValue) {
-        this.defaultValue.set(defaultValue);
-    }
-
-    // Description
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public String getDescription() {
-        return description.get();
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setDescription(String description) {
-        this.description.set(description);
-    }
-
-    // Extension
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public Map<String, Object> getExtensions() {
-        return extensions.get();
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public boolean hasExtension(String name) {
-        return extensions.containsKey(name);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public Object getExtension(String name) {
-        return extensions.get(name);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setExtensions(Map<String, Object> extensions) {
-        this.extensions.set(extensions);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setExtension(String name, Object extension) {
-        extensions.set(name, extension);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void removeExtension(String name) {
-        extensions.remove(name);
-    }
-
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public static JsonOverlayFactory<ServerVariableImpl> factory = new JsonOverlayFactory<ServerVariableImpl>() {
-    @Override
-    public ServerVariableImpl create(String key, JsonNode json, JsonOverlay<?> parent) {
-        return isEmptyRecursive(parent, ServerVariableImpl.class) ? null : new ServerVariableImpl(key, json, parent);
-    }
-};
-
-}
