+++ reprezen-swagger-parser/src/main/java/com/reprezen/swaggerparser/jsonoverlay/gen/ImplGenerator.java
@@ -11,7 +11,6 @@ import org.apache.commons.lang3.StringUtils;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.base.Function;
 import com.google.common.collect.Lists;
-import com.reprezen.swaggerparser.impl3.SwaggerObjectImpl;
 import com.reprezen.swaggerparser.jsonoverlay.JsonOverlay;
 import com.reprezen.swaggerparser.jsonoverlay.JsonOverlayFactory;
 import com.reprezen.swaggerparser.jsonoverlay.coll.ListOverlay;
@@ -21,6 +20,7 @@ import com.reprezen.swaggerparser.jsonoverlay.coll.ValMapOverlay;
 import com.reprezen.swaggerparser.jsonoverlay.gen.SimpleJavaGenerator.Member;
 import com.reprezen.swaggerparser.jsonoverlay.gen.TypeData.Field;
 import com.reprezen.swaggerparser.jsonoverlay.gen.TypeData.Type;
+import com.reprezen.swaggerparser.ovl3.SwaggerObjectImpl;
 
 public class ImplGenerator extends TypeGenerator {
