+++ reprezen-swagger-parser/src/main/java/com/reprezen/swaggerparser/ovl3/CallbackImpl.java
@@ -1,0 +1,118 @@
+package com.reprezen.swaggerparser.ovl3;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import com.reprezen.swaggerparser.jsonoverlay.JsonOverlay;
+import com.reprezen.swaggerparser.jsonoverlay.JsonOverlayFactory;
+import com.reprezen.swaggerparser.jsonoverlay.coll.MapOverlay;
+import com.reprezen.swaggerparser.jsonoverlay.coll.ValMapOverlay;
+import com.reprezen.swaggerparser.jsonoverlay.std.AnyObjectOverlay;
+import com.reprezen.swaggerparser.model3.Callback;
+import com.reprezen.swaggerparser.model3.Path;
+import com.reprezen.swaggerparser.ovl3.PathImpl;
+import com.reprezen.swaggerparser.ovl3.SwaggerObjectImpl;
+import java.util.Map;
+import javax.annotation.Generated;
+
+public class CallbackImpl extends SwaggerObjectImpl implements Callback {
+
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public CallbackImpl(String key, JsonNode json, JsonOverlay<?> parent) {
+        super(key, json, parent);
+    }
+
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public CallbackImpl(String key, JsonOverlay<?> parent) {
+        super(key, parent);
+    }
+
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    private MapOverlay<PathImpl> callbackPaths = registerField("", "callbackPaths", "(?!x-).*", new MapOverlay<PathImpl>("", this, PathImpl.factory, "(?!x-).*"));
+
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    private ValMapOverlay<Object,AnyObjectOverlay> extensions = registerField("", "extensions", "x-.*", new ValMapOverlay<Object, AnyObjectOverlay>("", this, AnyObjectOverlay.factory, "x-.*"));
+
+    // CallbackPath
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public Map<String, ? extends Path> getCallbackPaths() {
+        return callbackPaths.get();
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public boolean hasCallbackPath(String expression) {
+        return callbackPaths.containsKey(expression);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public Path getCallbackPath(String expression) {
+        return callbackPaths.get(expression);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void setCallbackPaths(Map<String, ? extends Path> callbackPaths) {
+        @SuppressWarnings("unchecked")
+            Map<String,PathImpl> implCallbackPaths = (Map<String, PathImpl>) callbackPaths;
+            this.callbackPaths.set(implCallbackPaths);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void setCallbackPath(String expression, Path callbackPath) {
+        callbackPaths.set(expression, (PathImpl) callbackPath);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void removeCallbackPath(String expression) {
+        callbackPaths.remove(expression);
+    }
+
+    // Extension
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public Map<String, Object> getExtensions() {
+        return extensions.get();
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public boolean hasExtension(String name) {
+        return extensions.containsKey(name);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public Object getExtension(String name) {
+        return extensions.get(name);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void setExtensions(Map<String, Object> extensions) {
+        this.extensions.set(extensions);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void setExtension(String name, Object extension) {
+        extensions.set(name, extension);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void removeExtension(String name) {
+        extensions.remove(name);
+    }
+
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public static JsonOverlayFactory<CallbackImpl> factory = new JsonOverlayFactory<CallbackImpl>() {
+    @Override
+    public CallbackImpl create(String key, JsonNode json, JsonOverlay<?> parent) {
+        return isEmptyRecursive(parent, CallbackImpl.class) ? null : new CallbackImpl(key, json, parent);
+    }
+};
+
+}
