+++ reprezen-swagger-parser/src/main/java/com/reprezen/swaggerparser/ovl3/HeaderImpl.java
@@ -1,10 +1,10 @@
-package com.reprezen.swaggerparser.impl3;
+package com.reprezen.swaggerparser.ovl3;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.reprezen.swaggerparser.impl3.SwaggerObjectImpl;
 import com.reprezen.swaggerparser.jsonoverlay.JsonOverlay;
 import com.reprezen.swaggerparser.jsonoverlay.JsonOverlayFactory;
 import com.reprezen.swaggerparser.model3.Header;
+import com.reprezen.swaggerparser.ovl3.SwaggerObjectImpl;
 import javax.annotation.Generated;
 
 public class HeaderImpl extends ParameterImpl implements Header {
