+++ reprezen-swagger-parser/src/main/java/com/reprezen/swaggerparser/ovl3/LinkImpl.java
@@ -1,57 +1,144 @@
-package com.reprezen.swaggerparser.impl3;
+package com.reprezen.swaggerparser.ovl3;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.reprezen.swaggerparser.impl3.ServerVariableImpl;
-import com.reprezen.swaggerparser.impl3.SwaggerObjectImpl;
 import com.reprezen.swaggerparser.jsonoverlay.JsonOverlay;
 import com.reprezen.swaggerparser.jsonoverlay.JsonOverlayFactory;
-import com.reprezen.swaggerparser.jsonoverlay.coll.MapOverlay;
 import com.reprezen.swaggerparser.jsonoverlay.coll.ValMapOverlay;
 import com.reprezen.swaggerparser.jsonoverlay.std.AnyObjectOverlay;
 import com.reprezen.swaggerparser.jsonoverlay.std.StringOverlay;
-import com.reprezen.swaggerparser.model3.Server;
-import com.reprezen.swaggerparser.model3.ServerVariable;
+import com.reprezen.swaggerparser.model3.Link;
+import com.reprezen.swaggerparser.ovl3.SwaggerObjectImpl;
 import java.util.Map;
 import javax.annotation.Generated;
 
-public class ServerImpl extends SwaggerObjectImpl implements Server {
+public class LinkImpl extends SwaggerObjectImpl implements Link {
 
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public ServerImpl(String key, JsonNode json, JsonOverlay<?> parent) {
+    public LinkImpl(String key, JsonNode json, JsonOverlay<?> parent) {
         super(key, json, parent);
     }
 
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public ServerImpl(String key, JsonOverlay<?> parent) {
+    public LinkImpl(String key, JsonOverlay<?> parent) {
         super(key, parent);
     }
 
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    private StringOverlay url = registerField("url", "url", null, new StringOverlay("url", this));
+    private StringOverlay href = registerField("href", "href", null, new StringOverlay("href", this));
+
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    private StringOverlay operationId = registerField("operationId", "operationId", null, new StringOverlay("operationId", this));
+
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    private ValMapOverlay<String,StringOverlay> parameters = registerField("parameters", "parameters", null, new ValMapOverlay<String, StringOverlay>("parameters", this, StringOverlay.factory, null));
+
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    private ValMapOverlay<String,StringOverlay> headers = registerField("headers", "headers", null, new ValMapOverlay<String, StringOverlay>("headers", this, StringOverlay.factory, null));
 
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
     private StringOverlay description = registerField("description", "description", null, new StringOverlay("description", this));
 
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    private MapOverlay<ServerVariableImpl> serverVariables = registerField("variables", "serverVariables", "(?!x-)[a-zA-Z0-9\\._-]+", new MapOverlay<ServerVariableImpl>("variables", this, ServerVariableImpl.factory, "(?!x-)[a-zA-Z0-9\\._-]+"));
-
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    private ValMapOverlay<Object,AnyObjectOverlay> variablesExtensions = registerField("variables", "variablesExtensions", "x-.*", new ValMapOverlay<Object, AnyObjectOverlay>("variables", this, AnyObjectOverlay.factory, "x-.*"));
-
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
     private ValMapOverlay<Object,AnyObjectOverlay> extensions = registerField("", "extensions", "x-.*", new ValMapOverlay<Object, AnyObjectOverlay>("", this, AnyObjectOverlay.factory, "x-.*"));
 
-    // Url
+    // Href
     @Override
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public String getUrl() {
-        return url.get();
+    public String getHref() {
+        return href.get();
     }
 
     @Override
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setUrl(String url) {
-        this.url.set(url);
+    public void setHref(String href) {
+        this.href.set(href);
+    }
+
+    // OperationId
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public String getOperationId() {
+        return operationId.get();
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void setOperationId(String operationId) {
+        this.operationId.set(operationId);
+    }
+
+    // Parameter
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public Map<String, String> getParameters() {
+        return parameters.get();
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public boolean hasParameter(String name) {
+        return parameters.containsKey(name);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public String getParameter(String name) {
+        return parameters.get(name);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void setParameters(Map<String, String> parameters) {
+        this.parameters.set(parameters);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void setParameter(String name, String parameter) {
+        parameters.set(name, parameter);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void removeParameter(String name) {
+        parameters.remove(name);
+    }
+
+    // Header
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public Map<String, String> getHeaders() {
+        return headers.get();
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public boolean hasHeader(String name) {
+        return headers.containsKey(name);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public String getHeader(String name) {
+        return headers.get(name);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void setHeaders(Map<String, String> headers) {
+        this.headers.set(headers);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void setHeader(String name, String header) {
+        headers.set(name, header);
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void removeHeader(String name) {
+        headers.remove(name);
     }
 
     // Description
@@ -65,82 +152,6 @@ public class ServerImpl extends SwaggerObjectImpl implements Server {
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
     public void setDescription(String description) {
         this.description.set(description);
-    }
-
-    // ServerVariable
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public Map<String, ? extends ServerVariable> getServerVariables() {
-        return serverVariables.get();
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public boolean hasServerVariable(String name) {
-        return serverVariables.containsKey(name);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public ServerVariable getServerVariable(String name) {
-        return serverVariables.get(name);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setServerVariables(Map<String, ? extends ServerVariable> serverVariables) {
-        @SuppressWarnings("unchecked")
-            Map<String,ServerVariableImpl> implServerVariables = (Map<String, ServerVariableImpl>) serverVariables;
-            this.serverVariables.set(implServerVariables);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setServerVariable(String name, ServerVariable serverVariable) {
-        serverVariables.set(name, (ServerVariableImpl) serverVariable);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void removeServerVariable(String name) {
-        serverVariables.remove(name);
-    }
-
-    // VariablesExtension
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public Map<String, Object> getVariablesExtensions() {
-        return variablesExtensions.get();
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public boolean hasVariablesExtension(String name) {
-        return variablesExtensions.containsKey(name);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public Object getVariablesExtension(String name) {
-        return variablesExtensions.get(name);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setVariablesExtensions(Map<String, Object> variablesExtensions) {
-        this.variablesExtensions.set(variablesExtensions);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setVariablesExtension(String name, Object variablesExtension) {
-        variablesExtensions.set(name, variablesExtension);
-    }
-
-    @Override
-    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void removeVariablesExtension(String name) {
-        variablesExtensions.remove(name);
     }
 
     // Extension
@@ -181,10 +192,10 @@ public class ServerImpl extends SwaggerObjectImpl implements Server {
     }
 
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public static JsonOverlayFactory<ServerImpl> factory = new JsonOverlayFactory<ServerImpl>() {
+    public static JsonOverlayFactory<LinkImpl> factory = new JsonOverlayFactory<LinkImpl>() {
     @Override
-    public ServerImpl create(String key, JsonNode json, JsonOverlay<?> parent) {
-        return isEmptyRecursive(parent, ServerImpl.class) ? null : new ServerImpl(key, json, parent);
+    public LinkImpl create(String key, JsonNode json, JsonOverlay<?> parent) {
+        return isEmptyRecursive(parent, LinkImpl.class) ? null : new LinkImpl(key, json, parent);
     }
 };
