+++ reprezen-swagger-parser/src/main/java/com/reprezen/swaggerparser/ovl3/OperationImpl.java
@@ -1,14 +1,6 @@
-package com.reprezen.swaggerparser.impl3;
+package com.reprezen.swaggerparser.ovl3;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.reprezen.swaggerparser.impl3.CallbackImpl;
-import com.reprezen.swaggerparser.impl3.ExternalDocsImpl;
-import com.reprezen.swaggerparser.impl3.ParameterImpl;
-import com.reprezen.swaggerparser.impl3.RequestBodyImpl;
-import com.reprezen.swaggerparser.impl3.ResponseImpl;
-import com.reprezen.swaggerparser.impl3.SecurityRequirementImpl;
-import com.reprezen.swaggerparser.impl3.ServerImpl;
-import com.reprezen.swaggerparser.impl3.SwaggerObjectImpl;
 import com.reprezen.swaggerparser.jsonoverlay.JsonOverlay;
 import com.reprezen.swaggerparser.jsonoverlay.JsonOverlayFactory;
 import com.reprezen.swaggerparser.jsonoverlay.coll.ListOverlay;
@@ -26,6 +18,14 @@ import com.reprezen.swaggerparser.model3.RequestBody;
 import com.reprezen.swaggerparser.model3.Response;
 import com.reprezen.swaggerparser.model3.SecurityRequirement;
 import com.reprezen.swaggerparser.model3.Server;
+import com.reprezen.swaggerparser.ovl3.CallbackImpl;
+import com.reprezen.swaggerparser.ovl3.ExternalDocsImpl;
+import com.reprezen.swaggerparser.ovl3.ParameterImpl;
+import com.reprezen.swaggerparser.ovl3.RequestBodyImpl;
+import com.reprezen.swaggerparser.ovl3.ResponseImpl;
+import com.reprezen.swaggerparser.ovl3.SecurityRequirementImpl;
+import com.reprezen.swaggerparser.ovl3.ServerImpl;
+import com.reprezen.swaggerparser.ovl3.SwaggerObjectImpl;
 import java.util.Collection;
 import java.util.Map;
 import javax.annotation.Generated;
