+++ reprezen-swagger-parser/src/main/java/com/reprezen/swaggerparser/ovl3/TagImpl.java
@@ -1,25 +1,27 @@
-package com.reprezen.swaggerparser.impl3;
+package com.reprezen.swaggerparser.ovl3;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.reprezen.swaggerparser.impl3.SwaggerObjectImpl;
 import com.reprezen.swaggerparser.jsonoverlay.JsonOverlay;
 import com.reprezen.swaggerparser.jsonoverlay.JsonOverlayFactory;
 import com.reprezen.swaggerparser.jsonoverlay.coll.ValMapOverlay;
 import com.reprezen.swaggerparser.jsonoverlay.std.AnyObjectOverlay;
 import com.reprezen.swaggerparser.jsonoverlay.std.StringOverlay;
-import com.reprezen.swaggerparser.model3.License;
+import com.reprezen.swaggerparser.model3.ExternalDocs;
+import com.reprezen.swaggerparser.model3.Tag;
+import com.reprezen.swaggerparser.ovl3.ExternalDocsImpl;
+import com.reprezen.swaggerparser.ovl3.SwaggerObjectImpl;
 import java.util.Map;
 import javax.annotation.Generated;
 
-public class LicenseImpl extends SwaggerObjectImpl implements License {
+public class TagImpl extends SwaggerObjectImpl implements Tag {
 
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public LicenseImpl(String key, JsonNode json, JsonOverlay<?> parent) {
+    public TagImpl(String key, JsonNode json, JsonOverlay<?> parent) {
         super(key, json, parent);
     }
 
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public LicenseImpl(String key, JsonOverlay<?> parent) {
+    public TagImpl(String key, JsonOverlay<?> parent) {
         super(key, parent);
     }
 
@@ -27,7 +29,10 @@ public class LicenseImpl extends SwaggerObjectImpl implements License {
     private StringOverlay name = registerField("name", "name", null, new StringOverlay("name", this));
 
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    private StringOverlay url = registerField("url", "url", null, new StringOverlay("url", this));
+    private StringOverlay description = registerField("description", "description", null, new StringOverlay("description", this));
+
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    private ExternalDocsImpl externalDocs = registerField("externalDocs", "externalDocs", null, ExternalDocsImpl.factory.create("externalDocs", this));
 
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
     private ValMapOverlay<Object,AnyObjectOverlay> extensions = registerField("", "extensions", "x-.*", new ValMapOverlay<Object, AnyObjectOverlay>("", this, AnyObjectOverlay.factory, "x-.*"));
@@ -45,17 +50,30 @@ public class LicenseImpl extends SwaggerObjectImpl implements License {
         this.name.set(name);
     }
 
-    // Url
+    // Description
     @Override
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public String getUrl() {
-        return url.get();
+    public String getDescription() {
+        return description.get();
     }
 
     @Override
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public void setUrl(String url) {
-        this.url.set(url);
+    public void setDescription(String description) {
+        this.description.set(description);
+    }
+
+    // ExternalDocs
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public ExternalDocs getExternalDocs() {
+        return externalDocs;
+    }
+
+    @Override
+    @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
+    public void setExternalDocs(ExternalDocs externalDocs) {
+        this.externalDocs.set((ExternalDocsImpl) externalDocs);
     }
 
     // Extension
@@ -96,10 +114,10 @@ public class LicenseImpl extends SwaggerObjectImpl implements License {
     }
 
     @Generated("com.reprezen.swaggerparser.jsonoverlay.gen.CodeGenerator")
-    public static JsonOverlayFactory<LicenseImpl> factory = new JsonOverlayFactory<LicenseImpl>() {
+    public static JsonOverlayFactory<TagImpl> factory = new JsonOverlayFactory<TagImpl>() {
     @Override
-    public LicenseImpl create(String key, JsonNode json, JsonOverlay<?> parent) {
-        return isEmptyRecursive(parent, LicenseImpl.class) ? null : new LicenseImpl(key, json, parent);
+    public TagImpl create(String key, JsonNode json, JsonOverlay<?> parent) {
+        return isEmptyRecursive(parent, TagImpl.class) ? null : new TagImpl(key, json, parent);
     }
 };
